+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Crashes from process death
 - Crashes in widget handling
+- Android Auto Backup works again
 
 ### Changed